+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.1.1
 + Added resource links links
-
++ Fixed off by one issue
 ## v0.1.0
 + Refactored to use classes and add more type safety
 + Put input_hint outside of the inputbox instead of outside